+++ src/L2Data2Code/L2Data2Code.BaseGenerator/Configuration/IDataSorcesConfiguration.cs
@@ -1,6 +1,6 @@
 using L2Data2Code.SharedLib.Configuration;
 
-namespace L2Data2Code.SchemaReader.Configuration
+namespace L2Data2Code.BaseGenerator.Configuration
 {
     public interface IDataSorcesConfiguration : IBasicConfiguration<DataSourceConfiguration>
     {