+++ src/L2Data2Code/L2Data2Code.BaseGenerator/Configuration/ITemplatesConfiguration.cs
@@ -4,7 +4,7 @@ namespace L2Data2Code.BaseGenerator.Configuration
 {
     public interface ITemplatesConfiguration : IBasicConfiguration<TemplateConfiguration>
     {
-        bool HasToRemoveFolders(string key);
+        bool HasToRemoveFolders(string key, string varValue = null);
         string Resource(string key);
     }
 }