+++ src/L2Data2Code/L2Data2Code.BaseGenerator/Configuration/TemplatesConfiguration.cs
@@ -18,7 +18,16 @@ namespace L2Data2Code.BaseGenerator.Configuration
             };
         }
 
-        public bool HasToRemoveFolders(string key) => this[key]?.RemoveFolders ?? true;
+        public bool HasToRemoveFolders(string key, string varValue = null)
+        {
+            var result = this[key]?.RemoveFolders ?? true;
+            if (varValue != null)
+            {
+                result = bool.TryParse(varValue, out var varResultValue) ? varResultValue : result;
+            }
+            return result;
+        }
+
         public string Resource(string key) => this[key]?.ResourcesFolder ?? "General";
 
         private void SetVars()